+++ TSK.Tests/Steps/AutopsyAddDataSourceTests.cs
@@ -3,10 +3,10 @@ using System.Text;
 using System.Threading;
 using CSharpFunctionalExtensions;
 using Moq;
-using Reductech.Sequence.Connectors.TSK.Steps;
-using Reductech.Sequence.Core.ExternalProcesses;
+using Sequence.Connectors.TSK.Steps;
+using Sequence.Core.ExternalProcesses;
 
-namespace Reductech.Sequence.Connectors.TSK.Tests.Steps;
+namespace Sequence.Connectors.TSK.Tests.Steps;
 
 public partial class AutopsyAddDataSourceTests : StepTestBase<AutopsyAddDataSource, Unit>
 {