+++ TSK/Steps/AutopsyAddDataSource.cs
@@ -1,6 +1,6 @@
-﻿using Reductech.Sequence.Core.Internal.Errors;
+﻿using Sequence.Core.Internal.Errors;
 
-namespace Reductech.Sequence.Connectors.TSK.Steps;
+namespace Sequence.Connectors.TSK.Steps;
 
 /// <summary>
 /// Add a Data Source to an Autopsy Case